refactor!: fold Rask.Client into Rask.Wasm — IShare was never a third tier
Rask.Client existed to hold what "BOTH the WASM and Native hosts provide but the
Server host cannot" — its own csproj said so. With the native hosting model gone
there is no second host, so the package was an assembly, a DI tier, a test
project, a solution entry, a bundled DLL/XML pair in the Rask.Wasm nupkg and a
ProjectReference in six samples, all to carry one interface.

IShare is WASM-only by exactly the criterion Rask.Wasm.Browser already documents:
it needs transient user activation, which the Server's WebSocket round-trip
loses. It now sits beside IFullscreen and ISerial, registered by
AddWasmBrowserApis. AddClientBrowserApis and RaskClientBrowserApis are removed;
the WASM host registers two tiers instead of three.

BREAKING: Rask.Client.Browser.IShare / Share are now Rask.Wasm.Browser.IShare /
Share. Update the using. Pre-1.0, so no shim.

This also removes a wart the samples carried. Rask.Wasm referenced Rask.Client
with PrivateAssets="all", which kept it out of the nuspec deps but meant an
in-repo app referencing the host by ProjectReference did not get it — and a WASM
app needs it in the boot assembly manifest. Every WASM sample therefore carried a
ProjectReference it never compiled against, plus a comment explaining why. IShare
now ships inside Rask.Wasm.dll, which those samples already reference, so the
reference and the explanation both go.

ShareTests moved to Rask.Wasm.Tests, whose FakeJsRuntime was byte-identical to
the one it left behind. The TryAdd-fallback assertion moved with it into
RaskWasmBrowserApisTests, where the "registers nothing beyond the pinned set"
guard now covers IShare too.

Closes #823
