+++ samples/Rask.Example.Auth.WasmCookie/Rask.Example.Auth.WasmCookie.csproj
@@ -29,9 +29,6 @@
     <PackageReference Include="Microsoft.Extensions.DependencyInjection" />
     <ProjectReference Include="..\..\src\Rask.Core\Rask.Core.csproj"/>
     <ProjectReference Include="..\..\src\Rask.Html\Rask.Html.csproj"/>
-    <!-- Rask.Client (IShare) is a runtime dep of the WASM host; surface it in the boot manifest in-repo
-         (the host bundles it for package consumers via PrivateAssets="all"). -->
-    <ProjectReference Include="..\..\src\Rask.Client\Rask.Client.csproj"/>
     <ProjectReference Include="..\..\src\Rask.Wasm\Rask.Wasm.csproj"/>
     <ProjectReference Include="..\..\src\Rask.Generators\Rask.Generators.csproj"
                       OutputItemType="Analyzer"
