+++ samples/Rask.Example.Auth.WasmJwt/Rask.Example.Auth.WasmJwt.csproj
@@ -28,9 +28,6 @@
   <ItemGroup>
     <ProjectReference Include="..\..\src\Rask.Core\Rask.Core.csproj"/>
     <ProjectReference Include="..\..\src\Rask.Html\Rask.Html.csproj"/>
-    <!-- Rask.Client (IShare) is a runtime dep of the WASM host; surface it in the boot manifest in-repo
-         (the host bundles it for package consumers via PrivateAssets="all"). -->
-    <ProjectReference Include="..\..\src\Rask.Client\Rask.Client.csproj"/>
     <ProjectReference Include="..\..\src\Rask.Wasm\Rask.Wasm.csproj"/>
     <ProjectReference Include="..\..\src\Rask.Generators\Rask.Generators.csproj"
                       OutputItemType="Analyzer"
