+++ samples/Rask.Example.Shared/Features/Browser/ShareDemo.cs
@@ -7,7 +7,7 @@ namespace Rask.Example.Shared.Features;
 ///     <c>data-rask-share</c> attribute, so the click opens the OS share sheet from <b>any</b> host, the
 ///     Server included. The shared client fires <c>navigator.share</c> inside the gesture (no round-trip, so
 ///     the activation isn't lost). For a code-driven share on the in-process host, inject <c>IShare</c> from
-///     <c>Rask.Client.Browser</c>.
+///     <c>Rask.Wasm.Browser</c>.
 /// </summary>
 public sealed partial class ShareDemo : Component
 {
