+++ samples/Rask.Example.Site/Rask.Example.Site.csproj
@@ -35,11 +35,10 @@
 
   <ItemGroup>
     <ProjectReference Include="..\..\src\Rask.Wasm\Rask.Wasm.csproj"/>
-    <!-- Rask.Wasm marks Rask.Core / Rask.Client PrivateAssets="all" (kept out of its nuspec deps),
+    <!-- Rask.Wasm marks Rask.Core / Rask.Html PrivateAssets="all" (kept out of its nuspec deps),
          so an in-repo consumer referencing the host by ProjectReference must pull them in directly. -->
     <ProjectReference Include="..\..\src\Rask.Core\Rask.Core.csproj"/>
     <ProjectReference Include="..\..\src\Rask.Html\Rask.Html.csproj"/>
-    <ProjectReference Include="..\..\src\Rask.Client\Rask.Client.csproj"/>
     <PackageReference Include="Microsoft.JSInterop"/>
     <ProjectReference Include="..\..\src\Rask.Generators\Rask.Generators.csproj"
                       OutputItemType="Analyzer"
