+++ samples/Rask.Example.Wasm.Jobs/Rask.Example.Wasm.Jobs.csproj
@@ -36,11 +36,10 @@
 
   <ItemGroup>
     <ProjectReference Include="..\..\src\Rask.Wasm\Rask.Wasm.csproj"/>
-    <!-- Rask.Wasm marks Rask.Core / Rask.Client PrivateAssets="all" (kept out of its nuspec deps),
+    <!-- Rask.Wasm marks Rask.Core / Rask.Html PrivateAssets="all" (kept out of its nuspec deps),
          so an in-repo consumer referencing the host by ProjectReference must pull them in directly. -->
     <ProjectReference Include="..\..\src\Rask.Core\Rask.Core.csproj"/>
     <ProjectReference Include="..\..\src\Rask.Html\Rask.Html.csproj"/>
-    <ProjectReference Include="..\..\src\Rask.Client\Rask.Client.csproj"/>
     <!-- The point of the sample: the browser database, and the same job queue the server uses. -->
     <ProjectReference Include="..\..\src\Rask.SQLite.Browser\Rask.SQLite.Browser.csproj"/>
     <ProjectReference Include="..\..\src\Rask.Jobs\Rask.Jobs.csproj"/>
