--- src/Rask.Client/Rask.Client.csproj
@@ -1,36 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <!--
-    Rask.Client is the home for in-process CLIENT capabilities the WASM host provides but the Server host
-    cannot — device/browser APIs that need transient user activation or a live document, which the Server's
-    WebSocket round-trip loses. It sits between Rask.Core and the in-process host:
-      • Rask.Core.Browser — APIs every host (incl. Server) can serve over IJSRuntime.
-      • Rask.Client.Browser (here) — APIs the in-process WASM host can serve but Server cannot, because
-        they need a live user gesture the WebSocket round-trip loses (currently IShare).
-      • Rask.Wasm.Browser — the WASM-only APIs (Fullscreen, Serial, USB, …).
-
-    Like Rask.Core, this is IsPackable=false: its DLL is bundled into the Rask.Wasm package lib/ folder
-    rather than published as its own package. Its only dependency (Microsoft.JSInterop) is already
-    surfaced at the host package boundary.
-  -->
-  <PropertyGroup>
-    <TargetFramework>net10.0</TargetFramework>
-    <ImplicitUsings>enable</ImplicitUsings>
-    <Nullable>enable</Nullable>
-    <IsPackable>false</IsPackable>
-    <RootNamespace>Rask.Client</RootNamespace>
-  </PropertyGroup>
-
-
-  <ItemGroup>
-    <!-- ShareData (the shared payload) lives in Rask.Core.Browser — used by both the all-host declarative
-         Shareable in Rask.Html and this imperative IShare. Core is IsPackable=false and already bundled
-         into the host packages alongside Rask.Client. -->
-    <ProjectReference Include="..\Rask.Core\Rask.Core.csproj" PrivateAssets="all"/>
-    <ProjectReference Include="..\Rask.Html\Rask.Html.csproj" PrivateAssets="all"/>
-    <PackageReference Include="Microsoft.JSInterop"/>
-    <ProjectReference Include="..\Rask.Generators\Rask.Generators.csproj"
-                      OutputItemType="Analyzer" ReferenceOutputAssembly="false"/>
-  </ItemGroup>
-
-</Project>
