--- tests/Rask.Client.Tests/Rask.Client.Tests.csproj
@@ -1,38 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>net10.0</TargetFramework>
-    <ImplicitUsings>enable</ImplicitUsings>
-    <Nullable>enable</Nullable>
-    <IsPackable>false</IsPackable>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="coverlet.collector">
-      <PrivateAssets>all</PrivateAssets>
-      <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
-    </PackageReference>
-    <PackageReference Include="Microsoft.NET.Test.Sdk"/>
-    <PackageReference Include="xunit"/>
-    <PackageReference Include="xunit.runner.visualstudio">
-      <PrivateAssets>all</PrivateAssets>
-      <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
-    </PackageReference>
-  </ItemGroup>
-
-  <ItemGroup>
-    <Using Include="Xunit"/>
-    <Using Include="Rask.Client.Browser"/>
-    <Using Include="Rask.Core.Browser"/>
-  </ItemGroup>
-
-  <ItemGroup>
-    <!-- ShareData lives in Rask.Core.Browser; Rask.Client references Core privately, so name it directly. -->
-    <ProjectReference Include="..\..\src\Rask.Core\Rask.Core.csproj"/>
-    <ProjectReference Include="..\..\src\Rask.Html\Rask.Html.csproj"/>
-    <ProjectReference Include="..\..\src\Rask.Client\Rask.Client.csproj"/>
-    <ProjectReference Include="..\..\src\Rask.Generators\Rask.Generators.csproj"
-                      OutputItemType="Analyzer" ReferenceOutputAssembly="false"/>
-  </ItemGroup>
-
-</Project>
